Changed mapping - changed mapping to /api/v1/... for all api configurations - added overview and config controllers to separate some functions
