+++ src/main/java/dev/leonlatsch/oliviabackend/controller/AuthController.java
@@ -20,7 +20,7 @@ import static dev.leonlatsch.oliviabackend.util.ControllerUtils.createResponseEn
  */
 @RestController
 @CrossOrigin(origins = "*")
-@RequestMapping("/auth")
+@RequestMapping("/api/v1/auth")
 public class AuthController {
 
     @Autowired
