+++ src/main/java/dev/leonlatsch/oliviabackend/controller/UserController.java
@@ -22,7 +22,7 @@ import static dev.leonlatsch.oliviabackend.util.ControllerUtils.createResponseEn
  */
 @RestController
 @CrossOrigin(origins = "*")
-@RequestMapping("/user")
+@RequestMapping("/api/v1/user")
 public class UserController {
 
     @Autowired
